+++ docs/changelog.md
@@ -82,6 +82,7 @@ tad-compiler:
 GUI changes:
  * Improved the unsaved changes dialogs
  * Fixed set start ticks to cursor button (F9)
+ * Added a search bar to the editors (Ctrl+F)
 
 
 Version 0.1.1